// src/interfaces/Students.tsx
export interface Student {
    number?: number;
    _id?: number | string;
    name?: string;
    lastname?: string;
    username?: string;
    age?: string;
    curp?: string;
    email?: string;
    password?: string;
    phone?: string;
    address?: string;
    drug?: string;
    tutor?: string;
    file?: string;
    files?: string[];
    description?: string;
    startdate?: string;
    enddate?: string;
    status?: string;
}